Fix the ESP-IDF build: %u on a uint32_t; add diagnostics to the example config
breezy_spi.cpp formatted a volatile uint32_t with %u in the edge-histogram
line. uint32_t is `unsigned long` on the xtensa toolchain, so this is a
-Wformat violation, and the ESP-IDF build treats it as fatal. It is the only
format-checked call in either component - every other log site goes through
ESP_LOG*, which carries no format attribute - which is why this single line
was the whole failure. Cast to (unsigned), matching breezy_climate.cpp.

Verified in a container against the real ESP-IDF 5.5.5 xtensa toolchain: the
unpatched tree reproduces the failure exactly, the patched tree links clean
(RAM 28.3%, flash 54.3%).

The example config gains uptime, wifi_signal, heap-free and reset-reason
diagnostics. Without an uptime sensor there is no way to tell a node that
rebooted from one that stayed up while unreachable, and those have entirely
different causes; reset_reason further separates a brownout from a watchdog.

PROTOCOL_IR.md now says what the symbol letters abbreviate. The table gave
timings but never glossed the names, so LM reads as "long, medium" and LS as
"long dip, short hill" - both wrong, and both a natural first reading.

+++ esphome/components/breezy_spi/breezy_spi.cpp
@@ -108,7 +108,11 @@ void BreezySPIComponent::loop() {
     size_t hp = 0;
     for (int i = 0; i < 21 && hp < sizeof(hb) - 16; i++) {
       if (edge_hist_[i])
-        hp += snprintf(hb + hp, sizeof(hb) - hp, "%d:%u ", i * 32, edge_hist_[i]);
+        // Cast: uint32_t is `unsigned long` on this toolchain, so a bare %u is
+        // a -Wformat error, which the ESP-IDF build treats as fatal. Same idiom
+        // as the (unsigned) casts in breezy_climate.cpp.
+        hp += snprintf(hb + hp, sizeof(hb) - hp, "%d:%u ", i * 32,
+                       (unsigned) edge_hist_[i]);
     }
     ESP_LOGD(TAG, "edges: %s", hb);
   }
